feat(feature_websocket_api): fix things on some things that were a little broken on the api and made the block users on the api

+++ ApiMasterPCBuilder/src/main/java/es/iespuertodelacruz/ohjm/apimasterpcbuilder/infrastructure/adapter/secundary/persistence/BuildEntity.java
@@ -32,7 +32,7 @@ public class BuildEntity implements Serializable {
 	@Column(name="TOTAL_PRICE")
 	private double totalPrice;
 	@Column(name="DATE_OF_CREATION")
-	private BigInteger dateOfCreation;
+	private Long dateOfCreation;
 	//bi-directional many-to-one association to BuildComponentEntity
 	@OneToMany(mappedBy="build")
 	private List<BuildComponentEntity> buildsComponents;
@@ -49,11 +49,11 @@ public class BuildEntity implements Serializable {
 	public BuildEntity() {
 	}
 
-	public BigInteger getDateOfCreation() {
+	public Long getDateOfCreation() {
 		return dateOfCreation;
 	}
 
-	public void setDateOfCreation(BigInteger dateOfCreation) {
+	public void setDateOfCreation(Long dateOfCreation) {
 		this.dateOfCreation = dateOfCreation;
 	}
